+++ test/frontend/support/_bootstrap.js
@@ -1,6 +1,4 @@
 global.Notification = require('./_mockNotification');
 global.document = require('./_mockDocument');
 
-global.self = {};
-
 require('../setup')
