+++ test/frontend/support/_mount.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import { Provider } from 'react-redux';
+import { mount } from 'enzyme';
+import configureStore from 'redux-mock-store';
+import {
+  createMemorySource,
+  createHistory,
+  LocationProvider,
+  Router,
+} from '@reach/router';
+
+const mountRouter = (elem, url = '/', state = {}) => {
+  const mockStore = configureStore([]);
+  const source = createMemorySource(url);
+  const history = createHistory(source);
+  return mount(
+    <LocationProvider history={history}>
+      <Provider store={mockStore(state)}>
+        <Router>
+          {elem}
+        </Router>
+      </Provider>
+    </LocationProvider>
+  );
+};
+
+export default mountRouter;
